Prefer ptrdiff_t to int and avoid integer overflows. * fileio.c (make_temp_name): * font.c (font_parse_family_registry): Avoid integer overflow on string size calculation. * data.c (Faset): Likewise for byte index.

+++ src/data.c
@@ -2310,7 +2310,7 @@ bool-vector.  IDX starts at 0.  */)
 	{
 	  if (! SINGLE_BYTE_CHAR_P (c))
 	    {
-	      int i;
+	      ptrdiff_t i;
 
 	      for (i = SBYTES (array) - 1; i >= 0; i--)
 		if (SREF (array, i) >= 0x80)
